fix(FAT1-25): fixing loading the whole screen on any action taken
