+++ assets/translations/ar.json
@@ -154,7 +154,7 @@
   "price": "السعر",
   "noDescription": "لا يوجد وصف",
   "location": "2XVP+XC - الشيخ زايد.....",
-  "noItems": "لا توجد عناصر في السلة",
+  "noItems": "سلة التسوق الخاصة بك فارغة",
   "noToken": "أنت غير مسجل الدخول، اضغط على السهم للعودة لتسجيل الدخول",
   "cartPage": "عربة التسوق",
