+++ assets/translations/en.json
@@ -149,7 +149,7 @@
   "noName": "No Name",
   "noDescription": "No Description",
   "location": "2XVP+XC - Sheikh Zayed.....",
-  "noItems": "No items in the cart",
+  "noItems": "Your cart is empty",
   "noToken": "You are not logged in press to Arrow back to login",
   "cartPage": "Cart",
   "deleteCart": "Item removed successfully",
