+++ lib/core/constants/app_animations.dart
@@ -2,4 +2,5 @@ abstract class AppAnimations {
   static const String loadingMobile =
       "assets/animations/loading_mobile_animation.json";
   static const String emptyFile = "assets/animations/empty_file.json";
+  static const String emptyAnimation = "assets/animations/empty_animation.json";
 }
