+++ lib/domain/use_cases/cart/delete_cart_item_use_case/delete_cart_item_use_case.dart
@@ -2,11 +2,12 @@ import 'package:flowery_app/api/client/api_result.dart';
 import 'package:flowery_app/domain/entities/cart/delete_cart/delete_cart_item.dart';
 import 'package:flowery_app/domain/repositories/cart/delete_item_repositories/delete_item_repositories.dart';
 import 'package:injectable/injectable.dart';
+
 @injectable
-class DeleteCartItemUseCase{
+class DeleteCartItemUseCase {
   final DeleteItemRepositories _repositories;
   DeleteCartItemUseCase(this._repositories);
-  Future<Result<DeleteItemsEntity>> deleteCartItem(String productId) {
-    return _repositories.deleteCartItem(productId);
+  Future<Result<DeleteItemsEntity>> deleteCartItem(String productId) async {
+    return await _repositories.deleteCartItem(productId);
   }
 }
