--- lib/presentation/cart/view_model/cart_intent.dart
@@ -1,5 +1,0 @@
-sealed class CartIntent {}
-
-class LoadCartIntent extends CartIntent {
-  LoadCartIntent();
-}
