--- lib/presentation/cart/view_model/delete_cubit/delete_intent.dart
@@ -1,6 +1,0 @@
-sealed class DeleteCartIntent {}
-
-class DeleteCartItemIntent extends DeleteCartIntent {
-  final String productId;
-  DeleteCartItemIntent({required this.productId});
-}
