--- lib/presentation/cart/view_model/quantity_cubit/quantity_intent.dart
@@ -1,11 +1,0 @@
-sealed class QuantityIntent {}
-
-class UpdateCartQuantityIntent extends QuantityIntent {
-  final String productId;
-  final int quantity;
-
-  UpdateCartQuantityIntent({
-    required this.productId,
-    required this.quantity,
-  });
-}
