+++ lib/presentation/cart/views/widget/cart_app_bar.dart
@@ -1,0 +1,23 @@
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flowery_app/core/constants/app_text.dart';
+import 'package:flowery_app/utils/common_widgets/custom_app_bar.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class CartAppBar extends StatelessWidget implements PreferredSizeWidget {
+  const CartAppBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomAppBar(
+      titleWidget: RPadding(
+        padding: const EdgeInsets.symmetric(horizontal: 16),
+        child: Text(AppText.cartPage.tr()),
+      ),
+      isTitleWidget: true,
+    );
+  }
+
+  @override
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
+}
