+++ lib/presentation/cart/views_model/cart_cubit/cart_intent.dart
@@ -1,0 +1,14 @@
+sealed class CartIntent {}
+
+class LoadCartIntent extends CartIntent {}
+
+class UpdateCartQuantityIntent extends CartIntent {
+  final String productId;
+  final int quantity;
+  UpdateCartQuantityIntent({required this.productId, required this.quantity});
+}
+
+class DeleteCartItemIntent extends CartIntent {
+  final String productId;
+  DeleteCartItemIntent({required this.productId});
+}
