+++ lib/presentation/checkout/views/widgets/custom_payment_vise.dart
@@ -1,14 +1,13 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowery_app/core/constants/app_text.dart';
-import 'package:flowery_app/presentation/checkout_page/view_model/payment_cubit/payment_cubit.dart';
+import 'package:flowery_app/presentation/checkout/views_model/payment_cubit/payment_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:group_radio_button/group_radio_button.dart';
 
-
-class CustomPayment extends StatelessWidget {
-  const CustomPayment({super.key});
+class CustomPaymentVise extends StatelessWidget {
+  const CustomPaymentVise({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -24,8 +23,8 @@ class CustomPayment extends StatelessWidget {
             border: Border.all(color: theme.colorScheme.shadow),
           ),
           child: RadioButton(
-            description: AppText.paymentCash.tr(),
-            value: AppText.paymentCash.tr(),
+            description: AppText.paymentCard.tr(),
+            value: AppText.paymentCard.tr(),
             groupValue: method,
             onChanged: (val) {
               context.read<PaymentCubit>().selectMethod(val!);
@@ -37,4 +36,3 @@ class CustomPayment extends StatelessWidget {
     );
   }
 }
-
